+++ src/smt/assertions_manager.h
@@ -33,11 +33,58 @@ namespace smt {
  */
 class AssertionsManager
 {
+  /** The type of our internal map of defined functions */
+  typedef context::CDHashMap<Node, smt::DefinedFunction, NodeHashFunction>
+      DefinedFunctionMap;
+  /** The type of our internal assertion list */
+  typedef context::CDList<Node> AssertionList;
+  /** The type of our internal assignment set */
+  typedef context::CDHashSet<Node, NodeHashFunction> AssignmentSet;
  public:
    AssertionsManager(SmtEngine& smt, ResourceManager& rm);
   ~AssertionsManager();
+  /** Process a user push.*/
+  void notifyPush() ;
+  /**
+    * Process a user pop.  Clears out the non-context-dependent stuff in this
+    * SmtEnginePrivate.  Necessary to clear out our assertion vectors in case
+    * someone does a push-assert-pop without a check-sat. It also pops the
+    * last map of expression names from notifyPush.
+    */
+  void notifyPop();
+  Node applySubstitutions(TNode node);
+  /** Return true if given expression is a defined function. */
+  bool isDefinedFunction(Node func);
+
+  /**
+   * Simplify node "in" by expanding definitions and applying any
+   * substitutions learned from preprocessing.
+   */
+  Node simplify(TNode in);
+
+  /**
+   * Adds a formula to the current context.  Action here depends on
+   * the SimplificationMode (in the current Options scope); the
+   * formula might be pushed out to the propositional layer
+   * immediately, or it might be simplified and kept, or it might not
+   * even be simplified.
+   * The arguments isInput and isAssumption are used for bookkeeping for proofs.
+   * The argument maybeHasFv should be set to true if the assertion may have
+   * free variables. By construction, assertions from the smt2 parser are
+   * guaranteed not to have free variables. However, other cases such as
+   * assertions from the SyGuS parser may have free variables (say if the
+   * input contains an assert or define-fun-rec command).
+   *
+   * @param isAssumption If true, the formula is considered to be an assumption
+   * (this is used to distinguish assertions and assumptions)
+   */
+  void addFormula(
+    TNode n, bool inUnsatCore, bool inInput, bool isAssumption, bool maybeHasFv);
   /** finish init */
   void finishInit();
+  /** get ite skolem map */
+  IteSkolemMap& getIteSkolemMap() { return d_assertions.getIteSkolemMap(); }
+  ProcessAssertions* getProcessAssertions() { return &d_processor; }
  private:
   /** Reference to the SMT engine */
   SmtEngine& d_smt;
@@ -46,6 +93,9 @@ class AssertionsManager
   /** The assertions processor */
   ProcessAssertions d_proc;
   /** The abstract values utility */
+  std::unique_ptr<smt::AbstractValues> d_absValues;
+  /** An index of our defined functions */
+  DefinedFunctionMap* d_definedFunctions;
   /**
    * The assertion list (before any conversion) for supporting
    * getAssertions().  Only maintained if in incremental mode.
@@ -65,10 +115,23 @@ class AssertionsManager
    * one single assumption ~(a1 AND ... AND an).
    */
   std::vector<Node> d_assumptions;
+  /**
+   * List of items for which to retrieve values using getAssignment().
+   */
+  AssignmentSet* d_assignments;
   /*
    * Whether we did a global negation of the formula.
    */
   bool d_globalNegation;
+  /** Assertions in the preprocessing pipeline */
+  AssertionPipeline d_assertions;
+  /** Whether any assertions have been processed */
+  CDO<bool> d_assertionsProcessed;
+  /** Instance of the ITE remover */
+  RemoveTermFormulas d_iteRemover;
+
+  /** The preprocessing pass context */
+  std::unique_ptr<PreprocessingPassContext> d_preprocessingPassContext;
 };
 
 }  // namespace smt
